# MakeWeBet.API/OpenApiEndPoint.yaml
---
openapi: 3.0.1
info:
  title: MakeWeBet.API
  version: '1.0'
paths:
  "/api/Customer/barcode":
    get:
      tags:
      - Customer
      parameters:
      - name: customerId
        in: query
        style: form
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
  "/api/Customer/payment":
    put:
      tags:
      - Customer
      parameters:
      - name: customerId
        in: query
        style: form
        schema:
          type: integer
          format: int32
      - name: productId
        in: query
        style: form
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
  "/api/Product":
    get:
      tags:
      - Product
      parameters:
      - name: productId
        in: query
        style: form
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
    post:
      tags:
      - Product
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/ProductDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ProductDto"
      responses:
        '200':
          description: Success
  "/api/Store/api/product_registration":
    put:
      tags:
      - Store
      parameters:
      - name: productId
        in: query
        style: form
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
components:
  schemas:
    ProductDto:
      type: object
      properties:
        storeId:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        price:
          type: number
          format: double
      additionalProperties: false